Fix TestUtils style
