--- src/test/TestUtils.java
@@ -1,37 +1,0 @@
-/**
- * TestUtils.java
- * @author Erik Cole
- * @author Chris Waldon
- * @version 0.0.1
- *
- * Inspired by:
- * http://www.javaworld.com/article/2073056/swing-gui-programming/automate-gui-tests-for-swing-applications.html
- */
-import java.awt.Component;
-import java.awt.Container;
-
-public class TestUtils {
-    	/**
-    	 * getChildNamed returns the child of the swing component with the provided
-    	 * name.
-    	 * @param parent - The component to search within
-    	 * @param childName - The name of the child we are searching for
-    	 * @return the component that has the provided name, or null if nonexistent
-    	 */
-	public static Component getChildNamed(Component parent, String childName) {
-		if (childName.equals(parent.getName())) {
-    			return parent;
-		}
-		if (parent instanceof Container) {
-			Component[] children = ((Container)parent).getComponents();
-
-			for (Component child : children) {
-				Component foundChild = getChildNamed(child, name);
-				if (foundChild != null) {
-    					return foundChild;
-				}
-			}
-		}
-		return null;
-	}
-}
